system/requirements.txt
numpy
cython
PyMuPDF==1.25.1
fastapi
torch==2.5.0
transformers
pillow
easyocr
torchvision
openai
azure-identity
dill
accelerate
timm
paddlepaddle
paddleocr
pdf2image
pytesseract
supervision==0.18.0
ultralytics==8.1.24
einops==0.8.0
langchain_chroma
langchain_openai
frontend
langchain_experimental
pdfplumber
python-docx
markdown2
